Function to print the prime number

+++ more_functions_nested_loops/100-prime_factor.c
@@ -12,7 +12,7 @@
  * Return: Always 0 (Success)
  */
 
-void prime_factor(int n)
+void prime_factor(long int n)
 {
 	int i = 2;
 	int vrai = 0;
